requirements.txt
Pillow == 7.2.0
art == 5.2
matplotlib == 3.2.2
numpy == 1.18.5
pandas == 1.0.5
torch == 1.6.0
torchvision == 0.7.0
